+++ addon/components/responsive-toolbar.ts
@@ -28,24 +28,49 @@ export default class ResponsiveToolbar extends Component {
       const observer = new ResizeObserver(this.handleResize.bind(this));
       observer.observe(element);
       this.toolbar = element;
+      return () => {
+        observer.disconnect();
+      };
     },
     { eager: false }
   );
 
   setUpMainToolbar = modifier(
     (element: HTMLElement) => {
+      const observer = new ResizeObserver(this.handleResize.bind(this));
+      observer.observe(element);
+      if (element.children.length) {
+        const childs = element.children;
+        for (const child of childs) {
+          observer.observe(child);
+        }
+      }
       this.main.reference = element;
       // Call handleResize to ensure the toolbar is correctly initialized
       this.handleResize();
+      return () => {
+        observer.disconnect();
+      };
     },
     { eager: false }
   );
 
   setUpSideToolbar = modifier(
     (element: HTMLElement) => {
+      const observer = new ResizeObserver(this.handleResize.bind(this));
+      observer.observe(element);
+      if (element.children.length) {
+        const childs = element.children;
+        for (const child of childs) {
+          observer.observe(child);
+        }
+      }
       this.side.reference = element;
       // Call handleResize to ensure the toolbar is correctly initialized
       this.handleResize();
+      return () => {
+        observer.disconnect();
+      };
     },
     { eager: false }
   );