+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Also observe responsive toolbar children for resize
+
 ### Dependencies
 - Bumps `@typescript-eslint/parser` from 5.55.0 to 5.58.0
 - Bumps `linkifyjs` from 4.1.0 to 4.1.1